[22-10-07] 그룹, 인증샷 수정시 이미지 삭제 오류 수정

+++ src/main/java/com/havit/finalbe/service/GroupService.java
@@ -242,7 +242,9 @@ public class GroupService {
         }
 
         Long originImage = groups.getImageId();
-        imageService.deleteImage(originImage);
+        if (!originImage.equals(groupRequestDto.getImageId())) {
+            imageService.deleteImage(originImage);
+        }
         groups.update(groupRequestDto);
 
         if (null == groupRequestDto.getGroupTag()) {
